move usePlanType to shared location
